Refactor: `sidebars` -> `ui` name change for reducers and other feedback changes.
