+++ public/js/actions/index.js
@@ -6,9 +6,9 @@ const sources = require("./sources");
 const tabs = require("./tabs");
 const pause = require("./pause");
 const navigation = require("./navigation");
-const sidebars = require("./sidebars");
+const ui = require("./ui");
 
 module.exports = (Object.assign(
   navigation,
-  breakpoints, eventListeners, sources, tabs, pause, sidebars
+  breakpoints, eventListeners, sources, tabs, pause, ui
 ) : typeof breakpoints);
