+++ public/js/reducers/tests/ui.js
@@ -2,7 +2,7 @@
 declare var describe: (name: string, func: () => void) => void;
 declare var it: (desc: string, func: () => void) => void;
 
-const { State, update } = require("../sidebars");
+const { State, update } = require("../ui");
 const expect = require("expect.js");
 const C = require("../../constants");
